more
See previous two commits.

Changes to DKBSABrowseDialog were meant to make it into the first of those, but since I forgot, I'll describe them here. In addition to adjusting the thing to match some changes to DKGameFilePicker, these changes should also fix it not opening to the right folder (the path stem) when not given an initially selected file.

+++ native/src/DovahKitQtCustomWidgets/widgets/widget-dialogs/DKFormDestructionStageDialog.cpp
@@ -16,7 +16,7 @@ DKFormDestructionStageDialog::DKFormDestructionStageDialog(QWidget* parent) : QD
 
 DKFormDestructionStageDialog::DestructionStage DKFormDestructionStageDialog::value() const {
    DestructionStage out = {
-      .health_percent   = (uint16_t)this->ui.healthPerc->value(),
+      .health_percent   = (uint8_t)this->ui.healthPerc->value(),
       .damage_stage     = (uint8_t)this->ui.damageStage->value(),
       .self_damage_rate = (uint32_t)this->ui.selfDamage->value(),
       .explosion        = this->ui.explosion->formStub(),
